Extract classes and functions from analyse_barcodes.py utility into new library module.
